eventing/processing-pipelines/bigquery/chart-creator/python/requirements.txt
Flask
gunicorn
pandas
matplotlib==3.2.2
cloudevents==1.2.0
google.cloud.bigquery
google.cloud.storage
